Adding-Migration

+++ MoneyFellows.Application/Features/Products/Commands/CreateProduct/CreateProductCommand.cs
@@ -5,7 +5,7 @@ namespace MoneyFellows.Application.Features.Products.Commands.CreateProduct
 {
     public class CreateProductCommand : IRequest<Response<Unit>>
     {
-        public CreateProductCommand(string name, string description, byte[] image, decimal price, string merchant)
+        public CreateProductCommand(string name, string description, byte[] image, double price, string merchant)
         {
             Name = name;
             Description = description;
@@ -19,7 +19,7 @@ namespace MoneyFellows.Application.Features.Products.Commands.CreateProduct
         public string Name { get; set; }
         public string Description { get; set; }
         public byte[] Image { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public string Merchant { get; set; }
     }
 }
