+++ MoneyFellows.Application/Features/Products/Commands/UpdateProduct/UpdateProductCommand.cs
@@ -5,7 +5,7 @@ namespace MoneyFellows.Application.Features.Products.Commands.UpdateProduct
 {
     public class UpdateProductCommand : IRequest<Response<bool>>
     {
-        public UpdateProductCommand(string name, string description, byte[] image, decimal price, string merchant, Guid id)
+        public UpdateProductCommand(string name, string description, byte[] image, double price, string merchant, Guid id)
         {
             Name = name;
             Description = description;
@@ -21,7 +21,7 @@ namespace MoneyFellows.Application.Features.Products.Commands.UpdateProduct
         public string Name { get; set; }
         public string Description { get; set; }
         public byte[] Image { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public string Merchant { get; set; }
     }
 }
